For php 5.4
add AltoRouterTrait.php to use as trait 
and change AltoRouter class to use the AltoRouterTrait

You can now use altorouter as trait and as usual class. it normaly
doesn't break anything.
