+++ examples/basic/index.php
@@ -1,8 +1,12 @@
 <?php
 
+require '../../AltoRouterTrait.php';
 require '../../AltoRouter.php';
 
 $router = new AltoRouter();
+
+var_dump($router);
+
 $router->setBasePath('/AltoRouter/examples/basic');
 $router->map('GET|POST','/', 'home#index', 'home');
 $router->map('GET','/users/', array('c' => 'UserController', 'a' => 'ListAction'));
